refactor(desktop): split monolithic main/index.ts into focused modules
Breaks down the 1,235-line Electron main process file into 15 focused modules:

- store.ts: electron-store configuration
- state.ts: shared global state (mainWindow, tray, isQuitting, etc.)
- app-url.ts: app URL getter with environment awareness
- media-permissions.ts: media access permission handlers
- window-injections.ts: desktop CSS & double-click injection
- window.ts: window creation, lifecycle, state persistence
- auth-session.ts: auth session preload/cache management
- tray.ts: system tray icon & menu
- menu.ts: application menu template
- updater.ts: auto-updater configuration & handlers
- ipc-handlers.ts: all IPC handler registration
- mcp-status.ts: MCP status broadcasting
- power-monitor.ts: power state tracking (suspend/resume)
- deep-links.ts: protocol handling & OAuth exchange

The new index.ts is a thin 99-line entry point that imports and wires modules together.
